--- src/auth/strategies/kakao.strategy.ts
@@ -1,22 +1,0 @@
-import { Injectable } from '@nestjs/common';
-import { ConfigService } from '@nestjs/config';
-import { PassportStrategy } from '@nestjs/passport';
-import { Profile, Strategy } from 'passport-kakao';
-import { AuthService } from '../auth.service';
-
-@Injectable()
-export class KakaoStrategy extends PassportStrategy(Strategy) {
-  constructor(private readonly configService: ConfigService, private readonly authService: AuthService) {
-    super({
-      clientID: configService.get<string>('KAKAO_RESTAPI_KEY'),
-      callbackURL: configService.get<string>('KAKAO_CALLBACK_URL'),
-      scopeSeparator: ['profile', 'email'],
-    });
-  }
-  // 향후 jwtService로 토큰 반환
-  async validate(accessToken: string, refreshToken: string, profile: Profile) {
-    console.log(accessToken);
-    console.log(refreshToken);
-    console.log(profile);
-  }
-}
